Initial commit
Training pong works.

Training time: ~45 minutes on 6 workers to get to 1/1 win/loss.

The memory usage seems way to high, but does not grow over time.

Added the ability for each worker to run multiple environments, running each each environment for 1 batch before cycling to the next. The idea is to get some of the benefits of more parallel works, without the additional resource costs.